Fixed Issues with .csproj file
